Fix memcmp4 not working issue at x64
Since -1u become 2^32 and size_t -1u is 2^64, it return false always until now.

Also, UniformBufferBindingDescriptor align not fit as it's struct size.
So we cannot simply use memcpy for this case.

+++ dali/internal/graphics/gles-impl/gles-graphics-program.cpp
@@ -68,7 +68,7 @@ inline bool memcmp4(A* a, B* b, uint32_t size)
   auto* pb = reinterpret_cast<const uint32_t*>(b);
   size >>= 2;
   while(size-- && *pa++ == *pb++);
-  return (-1u == size);
+  return (static_cast<size_t>(-1lu) == size);
 };
 
 /**
